Search for log files in the SubFolder
Changed the log file search so that it starts in the subfolder. This prevents finding log files from other zip files in other folders.
Updated the Cashin table so it blanks out values from 'this' row already reported in the previous row. Cleaner presentatation of data.
Dont report something that hasnt changed.