Added EJView, XmlParamView and AddKeyView to AP processing
Modified the application so a mix of AP and SP Views can be generated using the commandline. Added EJ, XmlParam and AddKey views.
Changed the IContext interface so all methods just pass the context.
Added the APLogLine project which is the list of APLines recognized.
Cleaned up the command line processing.