Adding SP Flat Line Parsing to the Tool
Added the ability for the parser to read the flat nwlog files produced when XTM runs on other vendor machines like DN. This is the first of many checkins.
Also added the ability for the parser to parse out native physical cash unit information. Currently the View does not support rendering this information but its there.
Also fixed a bug in the release.yml - it now creates the dist folder so the release should be created.